feat: handle node builtings better
This change is designed to handle the node builtins more effectively, since they were moved to `node:xxx`;

Fixes #110
