+++ src/lib/denoify.ts
@@ -58,13 +58,15 @@ export async function denoify(params: {
 
             const { denoifyImportExportStatement } = denoifyImportExportStatementFactory(
                 (() => {
-                    const { resolveNodeModuleToDenoModule } = resolveNodeModuleToDenoModuleFactory({
-                        "userProvidedPorts": denoifyParams?.ports ?? {},
-                        "dependencies": packageJsonParsed["dependencies"] ?? {},
-                        "devDependencies": packageJsonParsed["devDependencies"] ?? {},
-                        "log": console.log,
-                        getInstalledVersionPackageJson
-                    });
+                    const { resolveNodeModuleToDenoModule } = resolveNodeModuleToDenoModuleFactory(
+                        { getInstalledVersionPackageJson },
+                        {
+                            "userProvidedPorts": denoifyParams?.ports ?? {},
+                            "dependencies": packageJsonParsed["dependencies"] ?? {},
+                            "devDependencies": packageJsonParsed["devDependencies"] ?? {},
+                            "log": console.log
+                        }
+                    );
 
                     return id<Parameters<typeof denoifyImportExportStatementFactory>[0]>({
                         "getDestDirPath": ({ dirPath }) => path.join(denoDistPath, path.relative(srcDirPath, dirPath)),
