+++ src/tools/getFileContents.ts
@@ -1,0 +1,16 @@
+import * as fs from "fs/promises";
+
+export const getLocalFileContents = async (filePath: string) => {
+    try {
+        await fs.access(filePath);
+        return fs.readFile(filePath, { encoding: "utf8" });
+    } catch {
+        return undefined;
+    }
+};
+
+export const getRemoteFileContents = (url: string) =>
+    fetch(url).then(
+        res => (`${res.status}`.startsWith("2") ? res.text() : undefined),
+        () => undefined
+    );
